switch tool control to serial
