+++ arduino_drivers/abluo_base_tools/src/SERIAL_PROCESSING.h
@@ -1,12 +1,11 @@
-#ifndef I2C_PROCESSING_H
-#define I2C_PROCESSING_H
+#ifndef SERIAL_PROCESSING_H
+#define SERIAL_PROCESSING_H
 #include <Arduino.h>
-#include <Wire.h>
 
-namespace LibWireConstants
+namespace LibSerialConstants
 {
     const byte numChars = 32;
-    const byte numInputs = 2;
+    const byte numInputs = 10;
 }
 
 extern boolean newData;
@@ -15,7 +14,7 @@ extern char receivedChars[];
 
 static char *input;
 static char *token;
-extern void receiveEvent(int numChar);
+extern void recvWithEndMarker();
 extern void parseInput();
 
 #endif
